+++ src/core/execution/TrainExecution.ts
@@ -129,9 +129,8 @@ export class TrainExecution implements Execution {
       return;
     }
     this.hasCargo = true;
-    // Starts at 1: don't load tail engine
-    for (let i = 1; i < this.cars.length; i++) {
-      this.cars[i].setLoaded(true);
+    for (const car of this.cars) {
+      car.setLoaded(true);
     }
   }
 
@@ -150,13 +149,6 @@ export class TrainExecution implements Execution {
       targetUnit: this.destination.unit,
       trainType: TrainType.Engine,
     });
-    // Tail is also an engine, just for cosmetics
-    this.cars.push(
-      this.player.buildUnit(UnitType.Train, tile, {
-        targetUnit: this.destination.unit,
-        trainType: TrainType.TailEngine,
-      }),
-    );
     for (let i = 0; i < this.numCars; i++) {
       this.cars.push(
         this.player.buildUnit(UnitType.Train, tile, {